2023.01.27 13시
주사위 애니메이션 제작
주사위 UI 구현
씬 배치 미세조정

+++ Assets/Script/Timer.cs
@@ -14,6 +14,6 @@ public class Timer : MonoBehaviour
 
     void Update()
     {
-        timer.text = "Time : " + GameManager.time.ToString() + " sec";
+        timer.text = "Time : " + ((int)GameManager.time).ToString() + " sec";
     }
 }
